Refactor module names with Otel flavoring
