+++ src/data.rs
@@ -1,2 +1,4 @@
+/// Diagnostic bundle types and data structures
+pub mod diagnostic;
 /// Elasticsearch data types and structures
 pub mod elasticsearch;
