+++ src/data/diagnostic/kibana.rs
