+++ src/data/diagnostic/logstash.rs
