+++ src/lib.rs
@@ -2,14 +2,14 @@
 pub mod data;
 /// Environment variables
 pub mod env;
+/// Exports data to various destinations
+pub mod exporter;
 /// Manage saving and loading hosts from a YAML file
 pub mod host;
-/// Receive data from various sources
-pub mod input;
-/// Exports data to various destinations
-pub mod output;
 /// Data transformation and processing logic
 pub mod processor;
+/// Receive data from various sources
+pub mod receiver;
 /// Send pre-built assets (index templates, etc) to Elasticsearch
 pub mod setup;
 /// Classify an input string as a type of univeral resource identifier (URI)
